backend/requirements.txt
fastapi
uvicorn
rembg[cpu]
python-multipart
pillow
onnxruntime
